requirements.txt
Pillow==8.4.0
pycryptodome==3.10.1